+++ ReleaseNotes.md
@@ -1,8 +1,10 @@
 # Version 1.0.0.5
 20th January 2022
 
+- Introduced TradeBuilder to simplify trade request creation
 - Fixed using delegates for "GET" operations
 - Updated tests to take into account default permissions for creating assets
+
 
 # Version 1.0.0.4
 24th March 2021